requirements.txt
python-telegram-bot
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
